papr: add pipeline for deploying customized Jenkins
This is the first step towards getting PAPR into an OpenShift instance.
This only sets up the *infrastructure* for PAPR, not PAPR itself. See
the `README.md` for more information and how to get started.

I'd like to have the actual job that executes the PAPR test runs be a
Jenkins Pipeline. I'd rather avoid having to create JJB YAML files if
possible.

The next step is now to modify PAPR itself to run as a pipeline and,
more importantly, to run containerized tests in pods rather than
assuming access to `docker.sock`.