papr/jenkins: GHPRB fixes and webhook enhancements
Fix GHPRB configuration so that we explicitly refer to the GitHub token
credentials ID. I initially thought this wasn't needed because it would
auto-detect it. But it only does this if you visit the Configurations
page at least once.

Add an auto-generated parameter for a shared webhook secret. This is
important because, unlike OpenShift-generated webhooks, the URL itself
doesn't contain a secret, so anyone could POST anything to us.

On init, a Groovy script will pick up the secret and inject it into the
GHPRB configurations.