papr: add slave and and job builder buildconfigs
This almost completes the picture for infra setup. We create a Jenkins
slave image that we'll use for all executions, and also create a
job-builder bc which is used to define the GHPRB gate jobs.

The next step is to connect those jobs to start a PAPR pipeline where
we'll actually run the PAPR executable.