jenkins/jjb: rename ghprb-gate to papr and flesh out
Let's keep all jobs related to PAPR in this new `papr.yaml` file.

This also adds a new `papr/papr-trigger.py` which converts the event
into a new Job. The next step now is to start adapting PAPR as follow:

- parse YAML
- to start, ignore suites that are not 'container' envs
- for each container testsuite, migrate from `docker` to `oc` to create
  a new Job and monitor it

I'm thinking of doing these modifications in a new `openshift` branch of
PAPR to avoid breakage in the current infra.